Clearn import in TypeChecker.ts
